+++ react/ProductSummaryCustom.tsx
@@ -163,7 +163,7 @@ function ProductSummaryCustom({
         to: href,
         onClick: autocompleteSummary ? actionOnClick : undefined,
         onClickCapture: autocompleteSummary ? undefined : actionOnClick,
-        'aria-label': `View product details for ${
+        'aria-label': `${
           product?.productName || 'product'
         }`,
       }
@@ -177,7 +177,7 @@ function ProductSummaryCustom({
         query,
         onClick: autocompleteSummary ? actionOnClick : undefined,
         onClickCapture: autocompleteSummary ? undefined : actionOnClick,
-        'aria-label': `View product details for ${
+        'aria-label': `${
           product?.productName || 'product'
         }`,
       }