+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
  ## [Unreleased]
 
+### Added
+
+- Implements support for Product Summary components to render optional props.
+
 ## [2.91.0] - 2025-11-05
 
 ### Added